Importing Coverity Scan modeling file circa September 6, 2021
The imported contents was last modified sometime before the above date
(unknown to me). The currently deployed Coverity modeling file is also
different, due to ongoing Squid Project experiments.

I assume that the "Modeling file for Squid" title that Coverity shows
when rendering that file is not a part of the modeling file itself. I
have not included that part in this import.

Captured at https://scan.coverity.com/projects/squid/model_file